split chain filter buttons
